requirements.txt
playwright==1.44.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
python-dotenv==1.0.1
